Fix products-service data handler

+++ products-service/src/main/java/pl/bielamarcin/productsservice/service/ProductService.java
@@ -1,10 +1,11 @@
 package pl.bielamarcin.productsservice.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import pl.bielamarcin.productsservice.dto.ProductDTO;
+import pl.bielamarcin.productsservice.mapper.ProductMapper;
 import pl.bielamarcin.productsservice.repository.ProductRepository;
 import pl.bielamarcin.productsservice.model.Product;
-import pl.bielamarcin.productsservice.repository.ProductRepository;
+import pl.bielamarcin.productsservice.exception.ProductNotFoundException;
 
 import java.util.List;
 
@@ -12,34 +13,52 @@ import java.util.List;
 public class ProductService {
     private final ProductRepository productRepository;
 
-    @Autowired
-    public ProductService(ProductRepository productRepository) {
+    private final ProductMapper productMapper;
+
+    public ProductService(ProductRepository productRepository, ProductMapper productMapper) {
         this.productRepository = productRepository;
+        this.productMapper = productMapper;
     }
 
-    public List<Product> getAllProducts() {
-        return productRepository.findAll();
+    public List<ProductDTO> getAllProducts() {
+        return productRepository.findAll().stream().map(productMapper::toDTO).toList();
     }
 
-    public Product getProductById(Long id) {
-        return productRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Product not found"));
-    }
-    public Product addProduct(Product product) {
-        return productRepository.save(product);
+    public ProductDTO getProductById(Long id) throws ProductNotFoundException{
+        Product product = productRepository.findById(id)
+                .orElseThrow(() -> new ProductNotFoundException("Product not found"));
+        return productMapper.toDTO(product);
     }
 
-    public Product updateProduct(Long id, Product updatedProduct) {
-        Product product = getProductById(id);
-        product.setName(updatedProduct.getName());
-        product.setDescription(updatedProduct.getDescription());
-        product.setPrice(updatedProduct.getPrice());
-        product.setQuantity(updatedProduct.getQuantity());
-        product.setCategory(updatedProduct.getCategory());
-        return productRepository.save(product);
+    public ProductDTO addProduct(ProductDTO productDTO) {
+        Product product = productRepository.save(productMapper.toEntity(productDTO));
+        return productMapper.toDTO(product);
+    }
+    public List<ProductDTO> addAllProducts(List<ProductDTO> productDTOs) {
+        List<Product> products = productDTOs.stream()
+                .map(productMapper::toEntity)
+                .toList();
+        List<Product> savedProducts = productRepository.saveAll(products);
+        return savedProducts.stream()
+                .map(productMapper::toDTO)
+                .toList();
     }
 
-    public void deleteProduct(Long id) {
+    public ProductDTO updateProduct(Long id, ProductDTO updatedProductDTO) throws ProductNotFoundException {
+        Product product = productRepository.findById(id)
+                .orElseThrow(() -> new ProductNotFoundException("Product not found"));
+        product.setName(updatedProductDTO.name());
+        product.setDescription(updatedProductDTO.description());
+        product.setPrice(updatedProductDTO.price());
+        product.setQuantity(updatedProductDTO.quantity());
+        product.setCategory(updatedProductDTO.category());
+        return productMapper.toDTO(productRepository.save(product));
+    }
+
+    public void deleteProduct(Long id) throws ProductNotFoundException {
+        if (!productRepository.existsById(id)) {
+            throw new ProductNotFoundException("Product not found");
+        }
         productRepository.deleteById(id);
     }
 }
